lookup-demo/requirements.txt
pymongo
python-dotenv
flask
flask-cors